+++ ui/cypress/constants/ApiReferencePage.js
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+
+export const url = '/main/apidocs';