+++ ui/src/Containers/Navigation/LeftNavigation.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
+import * as Icon from 'react-feather';
 import { connect } from 'react-redux';
-import { withRouter, NavLink as Link } from 'react-router-dom';
+import { withRouter, NavLink } from 'react-router-dom';
 import ReactRouterPropTypes from 'react-router-prop-types';
 import { createStructuredSelector } from 'reselect';
 import find from 'lodash/find';
@@ -9,9 +10,53 @@ import { selectors } from 'reducers';
 
 import { useTheme } from 'Containers/ThemeProvider';
 import NavigationPanel from './NavigationPanel';
-import ApiDocsNavigation from './ApiDocsNavigation';
-import LeftSideNavLinks, { navLinks } from './LeftSideNavLinks';
-import { getDarkModeLinkClassName } from './navHelpers';
+import { apidocsLink, navLinks, productdocsLink } from './LeftSideNavLinks';
+import { filterLinksByFeatureFlag } from './navHelpers';
+
+const iconBaseClass = 'h-4 w-4';
+const linkBaseClass = 'flex items-center';
+const textBaseClass =
+    'font-700 font-condensed leading-normal no-underline text-sm tracking-wide uppercase';
+const versionBaseClass = 'font-700 leading-normal px-3 py-1 text-xs text-center word-break-all';
+
+const InternalLink = ({
+    activeColorClass,
+    iconColorClass,
+    linkLayoutColorClass,
+    navLink,
+    showNavigationPanel,
+    textColorClass,
+}) => (
+    <NavLink
+        to={navLink.to}
+        activeClassName={activeColorClass}
+        onClick={showNavigationPanel(navLink)}
+        className={`${linkBaseClass} ${linkLayoutColorClass}`}
+        data-testid={navLink.data || navLink.text}
+    >
+        <div className="mr-2">
+            <navLink.Icon className={`${iconBaseClass} ${iconColorClass}`} />
+        </div>
+        <p className={`${textBaseClass} ${textColorClass}`}>{navLink.text}</p>
+    </NavLink>
+);
+
+const ExternalLink = ({ iconColorClass, linkLayoutColorClass, navLink, textColorClass }) => (
+    <a
+        href={navLink.to}
+        target="_blank"
+        rel="noopener noreferrer"
+        className={`${linkBaseClass} ${linkLayoutColorClass}`}
+    >
+        <div className="mr-2">
+            <navLink.Icon className={`${iconBaseClass} ${iconColorClass}`} />
+        </div>
+        <p className={`${textBaseClass} ${textColorClass}`}>{navLink.text}</p>
+        <div className="ml-2">
+            <Icon.ExternalLink className={`${iconBaseClass} ${iconColorClass}`} />
+        </div>
+    </a>
+);
 
 const versionString = (metadata) => {
     let result = `v${metadata.version}`;
@@ -21,50 +66,12 @@ const versionString = (metadata) => {
     return result;
 };
 
-const LeftNavigation = ({ location, metadata }) => {
+const LeftNavigation = ({ featureFlags, location, metadata }) => {
     const { isDarkMode } = useTheme();
 
     const [panelType, setPanelType] = useState(null);
     const [clickOnPanelItem, setClickOnPanelItem] = useState(false);
     const [selectedPanel, setSelectedPanel] = useState('');
-
-    const linkClassName = `flex font-condensed leading-normal font-700 text-primary-400 px-3 no-underline h-18 items-center border-b ${getDarkModeLinkClassName(
-        isDarkMode
-    )}`;
-
-    function getActiveClassName(navLink) {
-        const { pathname } = location;
-        const navText = navLink.text.toLowerCase();
-        const baseActiveClass = isDarkMode
-            ? 'text-primary-500 bg-primary-200 hover:bg-primary-200'
-            : 'bg-primary-700 hover:bg-primary-700 text-base-100';
-
-        if (
-            (pathname.includes('main/policies') ||
-                pathname.includes('main/integrations') ||
-                pathname.includes('main/access')) &&
-            navText === 'platform configuration'
-        ) {
-            return baseActiveClass;
-        }
-
-        if (navLink.to !== '') {
-            return baseActiveClass;
-        }
-        if (navLink.to === '') {
-            const baseFocusClass = isDarkMode
-                ? 'text-primary-500 bg-primary-200 hover:bg-primary-300'
-                : 'text-base-100 bg-base-800 hover:bg-base-800';
-            if (panelType && panelType === navLink.panelType) {
-                return baseFocusClass;
-            }
-            if (!panelType && clickOnPanelItem && selectedPanel === navText) {
-                return baseFocusClass;
-            }
-            return isDarkMode ? 'bg-base-100' : 'bg-primary-800';
-        }
-        return '';
-    }
 
     function closePanel(newClickOnPanelItem, newSelectedPanel) {
         return () => {
@@ -91,23 +98,6 @@ const LeftNavigation = ({ location, metadata }) => {
         };
     }
 
-    function renderLink(navLink) {
-        return (
-            <Link
-                to={navLink.to}
-                activeClassName={getActiveClassName(navLink)}
-                onClick={showNavigationPanel(navLink)}
-                className={linkClassName}
-                data-testid={navLink.data || navLink.text}
-            >
-                <div className="text-center pr-2">{navLink.renderIcon()}</div>
-                <div className={`${isDarkMode ? 'text-base-600' : 'text-base-100'}`}>
-                    {navLink.text}
-                </div>
-            </Link>
-        );
-    }
-
     function renderNavigationPanel() {
         if (!panelType) return '';
         return <NavigationPanel panelType={panelType} onClose={closePanel} />;
@@ -125,30 +115,100 @@ const LeftNavigation = ({ location, metadata }) => {
 
     useEffect(componentDidMount, []);
 
-    const darkModeClasses = isDarkMode
-        ? 'bg-base-100 border-t -mt-px border-r border-base-300'
+    const iconColorClass = 'text-primary-400';
+    const textColorClass = isDarkMode ? 'text-base-600' : 'text-base-100';
+    const versionColorClass = isDarkMode ? 'text-base-500' : 'text-primary-400';
+
+    const menuColorClass = isDarkMode
+        ? 'bg-base-100 border-base-300 border-r border-t -mt-px'
         : 'bg-primary-800';
+
+    const linkColorClass = isDarkMode
+        ? 'hover:bg-base-200 border-base-400'
+        : 'hover:bg-base-700 border-primary-900';
+
+    // Beware: react-router appends the following classes to active/focus link,
+    // but they have same specificity as any other Tailwind classes.
+    // Therefore, which class wins depends on order of rules in style element,
+    // not order of classes in attribute of link element.
+
+    const activeColorClass = isDarkMode
+        ? 'bg-primary-200 hover:bg-primary-200'
+        : 'bg-primary-700 hover:bg-primary-700';
+
+    const focusColorClass = isDarkMode
+        ? 'bg-primary-300 hover:bg-primary-300'
+        : 'bg-base-800 hover:bg-base-800';
+
+    function getActiveColorClass(navLink) {
+        const { pathname } = location;
+
+        if (navLink.to !== '') {
+            return activeColorClass;
+        }
+
+        if (navLink.paths && navLink.paths.some((path) => pathname.includes(path))) {
+            return activeColorClass;
+        }
+
+        if (panelType && panelType === navLink.panelType) {
+            return focusColorClass;
+        }
+
+        if (!panelType && clickOnPanelItem && selectedPanel === navLink.text.toLowerCase()) {
+            return focusColorClass;
+        }
+
+        return '';
+    }
+
+    // API Reference is not in navLinks array because:
+    // any extra vertical space is above it;
+    // therefore, it has border top;
+    // it has about half the height of the links above it.
+
     return (
         <>
             <div
-                className={`flex flex-col justify-between flex-none overflow-auto z-60 ignore-react-onclickoutside ${darkModeClasses}`}
+                className={`flex flex-col flex-none h-full overflow-auto z-60 ignore-react-onclickoutside ${menuColorClass}`}
             >
-                <nav className="left-navigation">
-                    <LeftSideNavLinks renderLink={renderLink} />
+                <nav className="flex flex-col flex-grow left-navigation">
+                    <ul className="flex flex-col h-full">
+                        {filterLinksByFeatureFlag(featureFlags, navLinks).map((navLink) => (
+                            <li key={navLink.text}>
+                                <InternalLink
+                                    activeColorClass={getActiveColorClass(navLink)}
+                                    iconColorClass={iconColorClass}
+                                    linkLayoutColorClass={`border-b h-18 px-3 ${linkColorClass}`}
+                                    navLink={navLink}
+                                    showNavigationPanel={showNavigationPanel}
+                                    textColorClass={textColorClass}
+                                />
+                            </li>
+                        ))}
+                        <li className="flex flex-col flex-grow justify-end">
+                            <InternalLink
+                                activeColorClass={getActiveColorClass(apidocsLink)}
+                                iconColorClass={iconColorClass}
+                                linkLayoutColorClass={`border-b border-t h-8 px-3 ${linkColorClass}`}
+                                navLink={apidocsLink}
+                                showNavigationPanel={showNavigationPanel}
+                                textColorClass={textColorClass}
+                            />
+                        </li>
+                        <li>
+                            <ExternalLink
+                                iconColorClass={iconColorClass}
+                                linkLayoutColorClass={`border-b h-8 pl-3 pr-2 ${linkColorClass}`}
+                                navLink={productdocsLink}
+                                textColorClass={textColorClass}
+                            />
+                        </li>
+                    </ul>
                 </nav>
-                <div
-                    className="flex flex-col h-full justify-end text-center text-xs font-700"
-                    data-testid="nav-footer"
-                >
-                    <ApiDocsNavigation onClick={closePanel()} />
-                    <span
-                        className={`left-navigation p-3 leading-normal ${
-                            isDarkMode ? 'text-base-500' : 'text-primary-400'
-                        } word-break-all`}
-                    >
-                        {versionString(metadata)}
-                    </span>
-                </div>
+                <p className={`left-navigation ${versionBaseClass} ${versionColorClass}`}>
+                    {versionString(metadata)}
+                </p>
             </div>
             {renderNavigationPanel()}
         </>
@@ -157,6 +217,12 @@ const LeftNavigation = ({ location, metadata }) => {
 
 LeftNavigation.propTypes = {
     location: ReactRouterPropTypes.location.isRequired,
+    featureFlags: PropTypes.arrayOf(
+        PropTypes.shape({
+            envVar: PropTypes.string.isRequired,
+            enabled: PropTypes.bool.isRequired,
+        })
+    ).isRequired,
     metadata: PropTypes.shape({
         version: PropTypes.string,
         releaseBuild: PropTypes.bool,
@@ -172,6 +238,7 @@ LeftNavigation.defaultProps = {
 };
 
 const mapStateToProps = createStructuredSelector({
+    featureFlags: selectors.getFeatureFlags,
     metadata: selectors.getMetadata,
 });