+++ ui/src/Containers/Navigation/navHelpers.js
@@ -1,3 +1,5 @@
+/* eslint-disable import/prefer-default-export */
+
 import { isBackendFeatureFlagEnabled } from 'utils/featureFlags';
 
 export function filterLinksByFeatureFlag(flagsToUse, navLinks, defaultVal = false) {
@@ -6,11 +8,3 @@ export function filterLinksByFeatureFlag(flagsToUse, navLinks, defaultVal = fals
         return isBackendFeatureFlagEnabled(flagsToUse, navLink.featureFlag, defaultVal);
     });
 }
-
-export const getDarkModeLinkClassName = (isDarkMode) =>
-    isDarkMode ? 'hover:bg-base-200 border-base-400' : 'border-primary-900 hover:bg-base-700';
-
-export default {
-    filterLinksByFeatureFlag,
-    getDarkModeLinkClassName,
-};